✨ Feat: HOC 구현 + 오류수정

+++ src/pages/api/dashboardEditApi.ts
@@ -2,8 +2,10 @@ import instance from '@/src/util/axios';
 
 export const getMembers = async (id: number, offset: number = 1) => {
   try {
-    const response = await instance.get(`/members?page=${offset}&size=4&dashboardId=${id}`);
-    return response.data;
+    if (localStorage.getItem('accessToken')) {
+      const response = await instance.get(`/members?page=${offset}&size=4&dashboardId=${id}`);
+      return response.data;
+    }
   } catch (error) {
     console.error(error);
   }
@@ -12,8 +14,10 @@ export const getMembers = async (id: number, offset: number = 1) => {
 
 export const getTotalMembers = async (id: number) => {
   try {
-    const response = await instance.get(`/members?page=1&size=20&dashboardId=${id}`);
-    return response.data;
+    if (localStorage.getItem('accessToken')) {
+      const response = await instance.get(`/members?page=1&size=20&dashboardId=${id}`);
+      return response.data;
+    }
   } catch (error) {
     console.error(error);
   }
@@ -22,8 +26,10 @@ export const getTotalMembers = async (id: number) => {
 
 export const getInvitees = async (id: number, offset: number = 1) => {
   try {
-    const response = await instance.get(`/dashboards/${id}/invitations?page=${offset}&size=4`);
-    return response.data;
+    if (localStorage.getItem('accessToken')) {
+      const response = await instance.get(`/dashboards/${id}/invitations?page=${offset}&size=4`);
+      return response.data;
+    }
   } catch (error) {
     console.error(error);
   }
@@ -32,8 +38,10 @@ export const getInvitees = async (id: number, offset: number = 1) => {
 
 export const getDashboard = async (id: number) => {
   try {
-    const response = await instance.get(`/dashboards/${id}`);
-    return response.data;
+    if (localStorage.getItem('accessToken')) {
+      const response = await instance.get(`/dashboards/${id}`);
+      return response.data;
+    }
   } catch (error) {
     console.error(error);
   }
@@ -45,8 +53,10 @@ export const handleDeleteMember = async (userId: number) => {
   if (!confirmDeletion) return;
 
   try {
-    await instance.delete(`/members/${userId}`);
-    alert('멤버를 삭제했습니다.');
+    if (localStorage.getItem('accessToken')) {
+      await instance.delete(`/members/${userId}`);
+      alert('멤버를 삭제했습니다.');
+    }
   } catch (error) {
     console.error(error);
   }
@@ -57,8 +67,10 @@ export const handleCancelInvitation = async (id: number, invitationId: number) =
   if (!confirmDeletion) return;
 
   try {
-    await instance.delete(`/dashboards/${id}/invitations/${invitationId}`);
-    alert('초대를 취소했습니다.');
+    if (localStorage.getItem('accessToken')) {
+      await instance.delete(`/dashboards/${id}/invitations/${invitationId}`);
+      alert('초대를 취소했습니다.');
+    }
   } catch (error) {
     console.error(error);
   }
@@ -66,8 +78,10 @@ export const handleCancelInvitation = async (id: number, invitationId: number) =
 
 export const getMyProfile = async () => {
   try {
-    const response = await instance.get(`/users/me`);
-    return response.data;
+    if (localStorage.getItem('accessToken')) {
+      const response = await instance.get(`/users/me`);
+      return response.data;
+    }
   } catch (error) {
     console.error(error);
   }
@@ -76,8 +90,10 @@ export const getMyProfile = async () => {
 
 export const getDashboardList = async (offset: number = 1) => {
   try {
-    const response = await instance.get(`/dashboards?navigationMethod=pagination&page=${offset}&size=10`);
-    return response.data;
+    if (localStorage.getItem('accessToken')) {
+      const response = await instance.get(`/dashboards?navigationMethod=pagination&page=${offset}&size=10`);
+      return response.data;
+    }
   } catch (error) {
     console.error(error);
   }
